Add `/browserconfig.xml` and tile images
By default, starting with IE11 on Windows 8.1, IE requests
`/browserconfig.xml` whenever a user hits the `pin site` button (same
as Safari for iOS does with the `apple-touch-icons` when the user hits
the share button).

The benefits of these file additions, include:

* reduces the number of 404s, especially as IE 11 and Windows 8.1 get
  more market share
* provides a starting point for dealing and customizing Windows tiles
* reminds users about the tile images, so that their site doesn't end
  up having as a tile the default IE11 logo, or (if the favicon is
  32x32px or bigger and not `.ico`) the site's favicon
* reduces the page size compare to using `<meta>` tags (the
  `/browserconfig.xml` isn't deliverd with every page, and it is only
  requested when the user clicks the `pin site button`)

Only two images are used instead of four, as:

* 558 x 558px (Large) - is used for the `Small`, `Medium`, and `Large`
  tiles, being automatically resized where necessary
* 558 x 270px (Wide)  - is used for the `Wide` tile as, in this case,
  resizing the 558 x 558px image just ends up looking weird

The two images used are 558 x 558px / 558 x 270px because that is the
recommended image size in order for the tiles to look good on a wide
range of devices.

Ref: http://msdn.microsoft.com/en-us/library/ie/dn455106.aspx